requirements.txt
python-dotenv>=1.0.0
python-telegram-bot>=20.6
SQLAlchemy>=1.4
alembic>=1.10.0
pydantic>=1.10
pymysql